Added hasLazyRelations flag to ColumnInfo and isLazy flag to RelationInfo;
Based on the RelationInfo flag, an "async" keyword is added for lazy relations;

The toLazy helper is also adjusted to take the hasLazyRelations flag into account;

This combination allows the laziness of relations to be controlled by the naming strategy, with the command line being merely the default. It's the naming strategy's responsibility to adjust both the original column and references to the tweaked name.